Waiting for services must be split
We can't necessarily even contact the other services until we've loaded
the URLs into the Directory.
